+++ extensions/moralis-auth copy.env
@@ -1,0 +1,6 @@
+LOCATION=us-central1
+MORALIS_API_KEY=projects/620009842012/secrets/ext-moralis-auth-MORALIS_API_KEY/versions/1
+SERVICE_ACCOUNT_EMAIL=firebase-adminsdk-gxp3c@sdxcrypto-algovera.iam.gserviceaccount.com
+SERVICE_ACCOUNT_PRIVATE_KEY=projects/620009842012/secrets/ext-moralis-auth-SERVICE_ACCOUNT_PRIVATE_KEY/versions/1
+SERVICE_ACCOUNT_PROJECT_ID=sdxcrypto-algovera
+WEBSITE_URI=https://app.algovera.ai
