+++ functions/package.json
@@ -11,17 +11,19 @@
   "engines": {
     "node": "16"
   },
-  "main": "index.js",
+  "main": "src/index.ts",
   "dependencies": {
     "coinbase-commerce-node": "^1.0.4",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "ethers": "^5.7.2",
+    "firebase": "^9.14.0",
     "firebase-admin": "^10.3.0",
-    "firebase-functions": "^4.0.2"
+    "firebase-functions": "^4.1.0"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.2.0"
   },
-  "private": true
+  "private": true,
+  "type": "module"
 }
